Rename components/TimeBarChart.vue to components/index/_shared/TimeBarChart.vue
