+++ components/index/CardsMonitoring/ConfirmedCasesNumber/Card.vue
@@ -49,7 +49,7 @@
 
 <script>
 import AppLink from '@/components/_shared/AppLink.vue'
-import TimeBarChart from '@/components/TimeBarChart.vue'
+import TimeBarChart from '@/components/index/_shared/TimeBarChart.vue'
 import Data from '@/data/data.json'
 import formatGraph from '@/utils/formatGraph'
