+++ components/index/CardsReference/DeathsByDeathDate/Card.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import TimeBarChart from '@/components/TimeBarChart.vue'
+import TimeBarChart from '@/components/index/_shared/TimeBarChart.vue'
 import deaths from '@/data/deaths.json'
 import calcDayBeforeRatio from '@/utils/calcDayBeforeRatio'
 import formatGraph from '@/utils/formatGraph'
