+++ components/index/CardsReference/PositiveNumberByDiagnosedDate/Card.vue
@@ -28,7 +28,7 @@
 </template>
 
 <script>
-import TimeBarChart from '@/components/TimeBarChart.vue'
+import TimeBarChart from '@/components/index/_shared/TimeBarChart.vue'
 import Data from '@/data/positive_by_diagnosed.json'
 import calcDayBeforeRatio from '@/utils/calcDayBeforeRatio'
 import formatGraph from '@/utils/formatGraph'
