+++ components/index/CardsReference/TelephoneAdvisoryReportsNumber/Card.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script>
-import TimeBarChart from '@/components/TimeBarChart.vue'
+import TimeBarChart from '@/components/index/_shared/TimeBarChart.vue'
 import Data from '@/data/data.json'
 import formatGraph from '@/utils/formatGraph'
