+++ components/index/_shared/TimeBarChart.vue
